Split out difference testing methods into a trait.
